[C++] move tests to subdirectory
- [R] don't need to delete test files anymore
